requirements.txt
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
platformdirs==4.3.7
virtualenv==20.30.0
vunit_hdl==4.7.0
